Add random intensity and previous intensity.
Intensity in a program stage can now be boolean false to use the previous stage intensity.
Intensity in a program stage can now be an object of {min:minDuty, max:maxDuty} to pick a random value. This random value is generated at each loop of the program, but not at stage repeats.